+++ src/mocks/handlers/cartItemList/cartItemListHandlers.ts
@@ -1,6 +1,6 @@
 import { http, HttpResponse } from "msw";
 
-import { CART_ITEMS_ENDPOINT } from "../../../api/config";
+import { CART_ITEMS_ENDPOINT } from "../../../apis/config";
 import productListData from "../productList/defaultData.json";
 import cartItemListData from "./defaultData.json";
