+++ src/mocks/handlers/productList/productListHandlers.ts
@@ -1,6 +1,6 @@
 import { http, HttpResponse } from "msw";
 
-import { PRODUCTS_ENDPOINT } from "../../../api/config";
+import { PRODUCTS_ENDPOINT } from "../../../apis/config";
 import { PRODUCT_LIST } from "../../../constants/productList";
 import { Product } from "../../../interfaces/Product";
 import productListData from "./defaultData.json";
